move block/sst into new format mod
