+++ slatedb/src/format/row_codec.rs
