+++ slatedb/src/garbage_collector/compacted_gc.rs
@@ -234,9 +234,9 @@ mod tests {
     use crate::compactions_store::{CompactionsStore, StoredCompactions};
     use crate::compactor_state::{Compaction, CompactionSpec, SourceId};
     use crate::db_state::{ManifestCore, SsTableId};
+    use crate::format::sst::SsTableFormat;
     use crate::manifest::store::StoredManifest;
     use crate::object_stores::ObjectStores;
-    use crate::sst::SsTableFormat;
     use crate::stats::StatRegistry;
     use crate::test_utils::build_test_sst;
     use object_store::{memory::InMemory, path::Path};
