+++ slatedb/src/lib.rs
@@ -45,11 +45,11 @@ pub use db_reader::DbReader;
 pub use db_snapshot::DbSnapshot;
 pub use db_transaction::DbTransaction;
 pub use error::{CloseReason, Error, ErrorKind};
+pub use format::sst::BlockTransformer;
 pub use garbage_collector::stats as garbage_collector_stats;
 pub use garbage_collector::GarbageCollectorBuilder;
 pub use merge_operator::{MergeOperator, MergeOperatorError};
 pub use rand::DbRand;
-pub use sst::BlockTransformer;
 pub use transaction_manager::IsolationLevel;
 pub use types::KeyValue;
 
@@ -62,6 +62,7 @@ pub mod compactor;
 pub mod config;
 pub mod db_cache;
 pub mod db_stats;
+pub mod format;
 pub mod manifest;
 pub mod seq_tracker;
 pub mod size_tiered_compaction;
@@ -70,8 +71,6 @@ pub mod stats;
 mod batch;
 mod batch_write;
 mod blob;
-mod block;
-mod block_iterator;
 #[cfg(any(test, feature = "bencher"))]
 mod bytes_generator;
 mod bytes_range;
@@ -115,10 +114,7 @@ mod rand;
 mod reader;
 mod retention_iterator;
 mod retrying_object_store;
-mod row_codec;
 mod sorted_run_iterator;
-mod sst;
-mod sst_iter;
 mod store_provider;
 mod tablestore;
 #[cfg(test)]
