+++ slatedb/src/reader.rs
@@ -4,11 +4,11 @@ use crate::clock::MonotonicClock;
 use crate::config::{DurabilityLevel, ReadOptions, ScanOptions};
 use crate::db_state::ManifestCore;
 use crate::db_stats::DbStats;
+use crate::format::sst_iter::{SstIterator, SstIteratorOptions};
 use crate::iter::{IterationOrder, KeyValueIterator};
 use crate::mem_table::{ImmutableMemtable, KVTable};
 use crate::oracle::Oracle;
 use crate::sorted_run_iterator::SortedRunIterator;
-use crate::sst_iter::{SstIterator, SstIteratorOptions};
 use crate::tablestore::TableStore;
 use crate::utils::get_now_for_read;
 use crate::utils::{build_concurrent, compute_max_parallel};
@@ -409,9 +409,9 @@ mod tests {
     use crate::batch::WriteBatch;
     use crate::clock::MonotonicClock;
     use crate::db_state::{SortedRun, SsTableHandle, SsTableId};
+    use crate::format::sst::SsTableFormat;
     use crate::object_stores::ObjectStores;
     use crate::oracle::DbReaderOracle;
-    use crate::sst::SsTableFormat;
     use crate::stats::StatRegistry;
     use crate::tablestore::TableStore;
     use object_store::{memory::InMemory, path::Path, ObjectStore};
