+++ slatedb/src/types.rs
@@ -184,7 +184,7 @@ impl ValueDeletable {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::row_codec::SstRowEntry;
+    use crate::format::row_codec::SstRowEntry;
     use rstest::rstest;
 
     // Tombstone encoding:
